SalomonAI_Organizado/services/ocr-processor/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pillow==11.0.0
python-multipart==0.0.9
boto3==1.35.83
pytesseract==0.3.13
numpy==2.1.3
